Fix SRT ingest stability, recording fidelity, and asset duration accuracy.
Initialize GStreamer before pydantic imports in the SRT child so srtsrc can accept callers (fixes OBS disconnect loops). Make the record path non-leaky with larger passthrough buffers, generate proper VOD HLS proxies, and derive asset duration from media rather than wall clock. Reset receiving stats when the encoder drops, add bus logging in the SRT child, and default SRT authentication to true.

+++ ingest_farm/models/__init__.py
@@ -3,7 +3,7 @@ from __future__ import annotations
 from datetime import datetime
 from typing import Any
 
-from sqlalchemy import JSON, Boolean, DateTime, ForeignKey, Integer, String, Text, func
+from sqlalchemy import JSON, BigInteger, Boolean, DateTime, ForeignKey, Integer, String, Text, func
 from sqlalchemy.orm import DeclarativeBase, Mapped, mapped_column, relationship
 
 from ingest_farm.schemas import new_id
@@ -40,7 +40,7 @@ class Recording(Base):
     status: Mapped[str] = mapped_column(String(32), default="recording")
     storage_path: Mapped[str] = mapped_column(Text, nullable=False)
     segment_count: Mapped[int] = mapped_column(Integer, default=0)
-    byte_size: Mapped[int] = mapped_column(Integer, default=0)
+    byte_size: Mapped[int] = mapped_column(BigInteger, default=0)
 
     channel: Mapped[Channel] = relationship(back_populates="recordings")
     asset: Mapped[Asset | None] = relationship(back_populates="recording", uselist=False)
